gridsight-repo/requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
opendssdirect.py>=0.8.4
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
pytest>=7.4
